Home: Fix <iframe> and <pre> overflow on mobile
There was a white gap to the right of the homepage on mobile,
because the <iframe> and <pre> for Browser and Node.js examples
was too wide.

When grid layout is in effect, the automatic overflow handling
doesn't work because the browser is not perceiving any natural
limit to the width of the box, instead being fairly weak in how
hard it tries to narrow any of the columns. It  will wrap where
possible for for things like <iframe> and none-linebreaking
<pre>, it can't make it any narrower. At that point, the overflow
would apply to the parent element for the entire grid, not the
individual columns.

To workaround this limitation, rather than rendering a single column
grid, just render them as natural block-level elements underneath
each other, thus allowing page and viewports widths to naturally
evoke the overflow handling one would expect through a scrollable
<iframe> and <pre>.