Build: Update to jekyll-theme-amethyst 0.2.1 (add 404 page)
The latest version removes the need for the "guides" fallback page,
which was only accessible when JavaScript is disabled. The latest
Amethyst version makes the menu accessible even when JS fails. It
was already accessible by keyboard or without CSS.

Also add a custom 404 page, matching the one from api.qunitjs.com.