Build: Exclude `h2.screen-reader-text` from Typesense
API pages like <https://api.qunitjs.com/assert/propEqual/>
have a lead paragraph with a hidden `<h2>Description</h2>`. It is
hidden by class="screen-reader-text".

The old jekyll-algolia got this right by accident because the hidden
heading is generated by a theme layout, not in the "content" portion
that it sees.

Hide it explicitly so that the lead paragraph isn't down ranked in
search results on par with other paragaphs under an `h2` heading.

Example, searching for "Compare":

== jekyll-algolia ==

Using autocomplete.js@0.37.1, as currently on qunitjs.com:

> * assert.propEqual()
>   Compare an object’s own properties using a strict comparison.

== typesense ==

Using the last scrape from before this commit:

> * assert.propEqual() > Examples
>   Compare the property values of two objects.

Hopefully this will fix it.